test: change the core event tests
